feat: C interface split
