+++ endpoint/CMakeLists.txt
@@ -26,7 +26,14 @@ endif()
 
 option(PRIVMX_BUILD_ENDPOINT_INTERFACE "Build endpoint interface" OFF)
 if(PRIVMX_BUILD_ENDPOINT_INTERFACE)
-    add_subdirectory(interface)
+    # add_subdirectory(interface)
+    add_subdirectory(interface/core)
+    add_subdirectory(interface/event)
+    add_subdirectory(interface/thread)
+    add_subdirectory(interface/store)
+    add_subdirectory(interface/kvdb)
+    add_subdirectory(interface/inbox)
+    add_subdirectory(interface/crypto)
 endif(PRIVMX_BUILD_ENDPOINT_INTERFACE)
 
 if(PRIVMX_BUILD_CLI AND PRIVMX_BUILD_ENDPOINT_ENDPOINT)
