+++ endpoint/interface/core/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointcoreinterface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointcoreinterface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointcoreinterface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointcoreinterface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointcoreinterface PUBLIC Poco::Foundation privmx privmxendpointcore)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointcoreinterface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointcoreinterface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointcoreinterface PRIVATE -Werror)
+endif()
