+++ endpoint/interface/crypto/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointcryptointerface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointcryptointerface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointcryptointerface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointcryptointerface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointcryptointerface PUBLIC Poco::Foundation privmx privmxendpointcoreinterface privmxendpointcore privmxendpointcrypto)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointcryptointerface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointcryptointerface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointcryptointerface PRIVATE -Werror)
+endif()
