+++ endpoint/interface/event/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointeventinterface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointeventinterface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointeventinterface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointeventinterface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointeventinterface PUBLIC Poco::Foundation privmx privmxendpointcoreinterface privmxendpointcore privmxendpointevent)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointeventinterface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointeventinterface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointeventinterface PRIVATE -Werror)
+endif()
