+++ endpoint/interface/inbox/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointinboxinterface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointinboxinterface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointinboxinterface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointinboxinterface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointinboxinterface PUBLIC Poco::Foundation privmx privmxendpointcoreinterface privmxendpointthreadinterface privmxendpointstoreinterface privmxendpointcore privmxendpointthread privmxendpointstore privmxendpointinbox)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointinboxinterface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointinboxinterface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointinboxinterface PRIVATE -Werror)
+endif()
