+++ endpoint/interface/kvdb/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointkvdbinterface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointkvdbinterface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointkvdbinterface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointkvdbinterface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointkvdbinterface PUBLIC Poco::Foundation privmx privmxendpointcoreinterface privmxendpointkvdb)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointkvdbinterface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointkvdbinterface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointkvdbinterface PRIVATE -Werror)
+endif()
