+++ endpoint/interface/store/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointstoreinterface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointstoreinterface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointstoreinterface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointstoreinterface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointstoreinterface PUBLIC Poco::Foundation privmx privmxendpointcoreinterface privmxendpointstore)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointstoreinterface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointstoreinterface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointstoreinterface PRIVATE -Werror)
+endif()
