+++ endpoint/interface/thread/CMakeLists.txt
@@ -1,0 +1,19 @@
+file(GLOB_RECURSE SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
+set(INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include ${CMAKE_CURRENT_BINARY_DIR}/include)
+
+if(BUILD_SHARED_LIBS)
+    add_library(privmxendpointthreadinterface SHARED ${SOURCES})
+else()
+    add_library(privmxendpointthreadinterface STATIC ${SOURCES})
+endif()
+target_include_directories(privmxendpointthreadinterface PUBLIC ${INCLUDE_DIRS})
+target_compile_options(privmxendpointthreadinterface PRIVATE -fPIC -Wall -Wextra -Wno-narrowing)
+target_link_libraries(privmxendpointthreadinterface PUBLIC Poco::Foundation privmx privmxendpointcoreinterface privmxendpointthread)
+if(BUILD_SHARED_LIBS)
+    set_target_properties(privmxendpointthreadinterface PROPERTIES POSITION_INDEPENDENT_CODE True)
+endif()
+install(TARGETS privmxendpointthreadinterface DESTINATION lib)
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/privmx DESTINATION include)
+if(PRIVMX_WERROR)
+    target_compile_options(privmxendpointthreadinterface PRIVATE -Werror)
+endif()
